+++ uplink/src/lib.rs
@@ -41,7 +41,6 @@
 //! [`port`]: base::AppConfig#structfield.port
 //! [`name`]: Action#structfield.name
 
-use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 use std::thread;
 use std::time::Duration;
@@ -49,7 +48,6 @@ use ::config::{Environment, File, FileFormat};
 use anyhow::Error;
 use flume::{bounded, Receiver, Sender};
 use log::{error, info};
-use structopt::StructOpt;
 
 pub mod console;
 pub mod base;
@@ -425,49 +423,6 @@ const DEFAULT_CONFIG: &str = r#"
     process_names = ["uplink"]
     update_period = 30
 "#;
-
-#[derive(StructOpt, Debug)]
-#[structopt(name = "uplink", about = "collect, batch, compress, publish")]
-pub struct CommandLine {
-    /// config file
-    #[structopt(short = "c", help = "Config file")]
-    pub config: Option<PathBuf>,
-    /// config file
-    #[structopt(short = "a", help = "Authentication file")]
-    pub auth: PathBuf,
-    /// log level (v: info, vv: debug, vvv: trace)
-    #[structopt(short = "v", long = "verbose", parse(from_occurrences))]
-    pub verbose: u8,
-    /// list of modules to log
-    #[structopt(short = "m", long = "modules")]
-    pub modules: Vec<String>,
-}
-
-pub fn initialize_logging(log_level: u8, log_modules: Vec<String>) {
-    let level = match log_level {
-        0 => "warn",
-        1 => "info",
-        2 => "debug",
-        _ => "trace",
-    };
-
-    let levels =
-        match log_modules.into_iter().reduce(|e, acc| format!("{e}={level},{acc}")) {
-            Some(f) => format!("{f}={level}"),
-            _ => format!("{level}"),
-        };
-
-    let builder = tracing_subscriber::fmt()
-        .pretty()
-        .with_line_number(false)
-        .with_file(false)
-        .with_thread_ids(false)
-        .with_thread_names(false)
-        .with_env_filter(levels)
-        .with_filter_reloading();
-
-    builder.try_init().expect("initialized subscriber succesfully");
-}
 
 fn parse_config(device_json: &str, config_toml: &str) -> Result<(Config, DeviceConfig), Error> {
     let mut config =
